services/mqtt_data_collector/requirements.txt
anyio==3.5.0
beanie==1.8.12
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
fastapi==0.72.0
Flask==2.0.2
Flask-MQTT==1.1.1
Flask-PyMongo==2.3.0
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
motor==2.5.1
multidict==5.2.0
mypy-extensions==0.4.3
paho-mqtt==1.6.1
pathspec==0.9.0
platformdirs==2.4.1
pydantic==1.9.0
pymongo==3.12.3
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
yapf==0.32.0
yarl==1.7.2
